separate sorting function into a new file
